Added a printallpkgs function, merged that functionality from boot and clean

+++ clean.sh
@@ -12,61 +12,51 @@ source "$(dirname "$(realpath "$0")")/lib/libmain.sh"
 source "$(dirname "$(realpath "$0")")/lib/libpackage.sh"
 source "$(dirname "$(realpath "$0")")/lib/libbuild.sh"
 
+remove_pkg() {
+    # Remove the given package file.
+    local file="$1"
+    local dry_run="$2"
+    message info "Removing outdated file '${file}'"
+    if ! "${dry_run}"; then rm "${file}"; fi
+}
+
 main() {
     # Remove all the old, outdated packages and source tarballs.
     local configdir="$1"
+    local dry_run="$2"
 
-    # Iterate through all of the packages.
-    local srctar srctarprefix pkg pkgprefix file
-    for dir in "${configdir}/src/"*/*; do
-        # Remove all outdated versions of this package.
+    # Generate a list of the current packages.
+    local dir pkgsdirs
+    pkgdirs=()
+    for dir in "${configdir}/src/"*; do
+        pkgdirs+=("$(basename "${dir}")")
+    done
+    local pkgs=($(printallpkgs "${configdir}" ${pkgdirs[@]}))
 
-        if [ ! -d "${dir}" ] || [ ! -f "${dir}/PKGBUILD" ]; then
-            # Bail early if this is not a valid package dir.
-            continue
-        fi
-        local pkgdir="$(printf "%s" "${dir}" | rev | cut -d'/' -f1-2 | rev)"
-
-        # We need a .SRCINFO - generate it if it does not exist.
-        walk "${configdir}" "rebuild" "srcinfo/${pkgdir}" || \
-            error 1 "Failed to generate a .SRCINFO for '${pkgdir}'!"
-        
-        # Find the expected names; iterate through the existing srctars and
-        # packages and remove all that match the package's name, but not the
-        # version.
-        srctar="$(pkgdirsrctar "${configdir}" "${pkgdir}")" || \
-            error 1 "Failed to generate the source tarball name for '${pkgdir}'!"
-        srctarprefix="$(printf "%s" "${srctar}" | rev | cut -d'-' -f3- | rev)"
-        for file in "${configdir}/srctar/${srctarprefix}"-[0-9]*-*"${SRCEXT}"; do
-            if [ "${file##*/}" != "${srctar}" ] && [ -f "${file}" ]; then
-                # Remove this; it is an outdated source tarball.
-                message info "Removing outdated file '${file}'"
-                rm "${file}"
+    # Iterate through all of the built packages, removing any that are not in
+    # the list of generated packages.
+    local pkgfile pkg
+    for pkgfile in "${configdir}/pkgs/"*; do
+        local in_list=false
+        for pkg in ${pkgs[@]}; do
+            if [ "${pkg}" == "${pkgfile}" ]; then
+                in_list=true
+                break
             fi
         done
-
-        # Iterate through the packages.
-        while IFS="-" read pkg; do
-            pkgprefix="$(printf "%s" "${pkg}" | rev | cut -d'-' -f4- | rev)"
-            for file in "${configdir}/pkgs/${pkgprefix}"-[0-9]*-*-*"${PKGEXT}"; do
-                if [ "${file##*/}" != "${pkg}" ] && [ -f "${file}" ]; then
-                    # Remove this; it is an outdated package.
-                    message info "Removing outdated file '${file}'"
-                    rm "${file}"
-                fi
-            done
-        done < <(pkgdirpackages "${configdir}" "${pkgdir}") || \
-            error 1 "Failed to generate a list of packages for '${pkgdir}'!"
+        "${in_list}" || remove_pkg "${pkgfile}" "${dry_run}"
     done
 }
 
 # Parse the arguments.
 CONFIGDIR="" # Set the initial config dir.
+DRY_RUN="false" # Wether or not to do a dry run.
 parseargs "$@" # Initial argument parse.
 # Manual argument parse.
 for arg in "$@"; do
     ignore_arg "${arg}" || \
     case "${arg}" in
+        -d|--dry-run) DRY_RUN=true;;
         *) if [ "${CONFIGDIR}" == "" ]; then
             CONFIGDIR="${arg}"
         else
@@ -76,4 +66,4 @@ for arg in "$@"; do
 done
 setup "${CONFIGDIR}"
 
-main "${CONFIGDIR}"
+main "${CONFIGDIR}" "${DRY_RUN}"
